ui: add connection and source UI components for Starrocks
This is a quick initial pass of updating the UI to support Starrocks. It is
almost completely just a copy-paste of the Clickhouse components, on the basis
that the connectivity and source configuration are the same. I have tried to
introduce the concept of a Catalog for Starrocks but this is done completely
blind, and so is probably incomplete.